Beta

+++ src/environment.rs
@@ -14,7 +14,7 @@ fn clock_impl(_args: &Vec<LiteralValue>) -> LiteralValue {
         .expect("Could not get system time")
         .as_millis();
     println!("{}", now);
-    LiteralValue::Number(now as f64 / 1000.0)
+    LiteralValue::Number(now as f32 / 1000.0)
 }
 fn get_globals() -> Rc<RefCell<HashMap<String, LiteralValue>>> {
     let mut env = HashMap::new();
